requirements.txt
sortedcontainers
TgCrypto
mysqlclient
telethon
pytelegrambotapi
cryptg
psycopg2


Flask
flask-wtf
flask-sqlalchemy
Flask-Migrate

python-dotenv